fixed int cast double failed error in partitioned and topic item (#36)

+++ lib/module/pulsar/pulsar_partitioned_topic_base.dart
@@ -3,9 +3,9 @@ class PulsarPartitionedTopicBaseResp {
   final int partitionNum;
   final double msgRateIn;
   final double msgRateOut;
-  final double msgInCounter;
-  final double msgOutCounter;
-  final double storageSize;
+  final int msgInCounter;
+  final int msgOutCounter;
+  final int storageSize;
 
   PulsarPartitionedTopicBaseResp(this.topicName, this.partitionNum, this.msgRateIn, this.msgRateOut, this.msgInCounter,
       this.msgOutCounter, this.storageSize);
